feat(core): prune mock and scenario entities
The initial state is grabbed from local storage when starting the
dashboard. This state can contain mocks/scenarios that are no longer
present in the active handler array passed to the setupDashboard
function.

We were relying on the entity adapter adding mocks in the fixed
order of the call stack. This does not work when adding new mocks while
local storage has already been populated. We fixed this by overriding
the entity id array with the ordered entity ids from the active handler
array.

ref: #141