feat: decouple from MSW internals
We used to be tightly coupled to the MSW internals for the sole purpose
of ensuring that msw handlers had access to the latest params and data.
This added a lot of complexity, decreased the maintainability. It also
does not work well with the new WebSocket implementation of MSW.

Decoupling also has a downside. Users now have to call `getParams()` and
`getData()` inside of their MSW resolver functions in order to get
access to up-to-date parameters and data. This is because we call the
handler initializing function only once now. At the time of writing, I'm
not sure how to tackle this problem differently. We could consider
creating a custom ESLint rule to prevent access of params/data outside
of the scope of the MSW resolver, if only to warn users about this.

BREAKING CHANGE: Check out the updated documentation to see how the
API's have changed. In short;
1. parameters and data are now obtained differently in your handlers.
   Refer to the updated examples in the readme.
2. you now have to call the functions `setupServer` and `setupWorker`
   exported from MSW. With this you have to initialize (dynamic) mocks
   and scenarios with the `setupHandlers` `setupDashboard` functions,
   exported from `dynamic-msw` and `@dynamic-msw/browser` respectively.
3. you now have to reset updates parameter/data updates using their
   `resetAll` function returned by `setupHandlers` `setupDashboard`